Separate tuning code from parallel.py
